[Input] Move keymap IDs into input/keymaps
