+++ cmake/treedata/common/input.txt
@@ -20,4 +20,3 @@ xbmc/input/keymaps/touch                        input/keymaps/touch
 xbmc/input/mouse                                input/mouse
 xbmc/input/mouse/generic                        input/mouse/generic
 xbmc/input/mouse/interfaces                     input/mouse/interfaces
-xbmc/input/remote                               input/remote
