+++ xbmc/dialogs/GUIDialogKeyboardGeneric.cpp
@@ -19,9 +19,9 @@
 #include "input/InputCodingTable.h"
 #include "input/actions/Action.h"
 #include "input/actions/ActionIDs.h"
-#include "input/keyboard/KeyIDs.h"
 #include "input/keyboard/KeyboardLayoutManager.h"
 #include "input/keyboard/XBMC_vkeys.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 #include "interfaces/AnnouncementManager.h"
 #include "messaging/ApplicationMessenger.h"
 #include "resources/LocalizeStrings.h"
