+++ xbmc/guilib/GUIBaseContainer.cpp
@@ -19,7 +19,7 @@
 #include "guilib/listproviders/IListProvider.h"
 #include "input/actions/Action.h"
 #include "input/actions/ActionIDs.h"
-#include "input/keyboard/KeyIDs.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 #include "input/mouse/MouseEvent.h"
 #include "settings/Settings.h"
 #include "settings/SettingsComponent.h"
