+++ xbmc/input/InputManager.cpp
@@ -21,13 +21,13 @@
 #include "input/actions/Action.h"
 #include "input/actions/ActionIDs.h"
 #include "input/keyboard/Key.h"
-#include "input/keyboard/KeyIDs.h"
 #include "input/keyboard/KeyboardEasterEgg.h"
 #include "input/keyboard/XBMC_vkeys.h"
 #include "input/keyboard/interfaces/IKeyboardDriverHandler.h"
 #include "input/keymaps/ButtonTranslator.h"
 #include "input/keymaps/KeymapEnvironment.h"
 #include "input/keymaps/joysticks/JoystickMapper.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 #include "input/keymaps/remote/CustomControllerTranslator.h"
 #include "input/keymaps/touch/TouchTranslator.h"
 #include "input/mouse/MouseTranslator.h"
