+++ xbmc/input/actions/Action.cpp
@@ -11,7 +11,7 @@
 #include "ActionIDs.h"
 #include "ActionTranslator.h"
 #include "input/keyboard/Key.h"
-#include "input/keyboard/KeyIDs.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 
 using namespace KODI;
 using namespace ACTION;
