+++ xbmc/input/keymaps/AppTranslator.cpp
@@ -8,7 +8,7 @@
 
 #include "AppTranslator.h"
 
-#include "input/keyboard/KeyIDs.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 #include "utils/Map.h"
 #include "utils/StringUtils.h"
 #include "utils/log.h"
