+++ xbmc/input/keymaps/ButtonTranslator.cpp
@@ -18,9 +18,9 @@
 #include "input/actions/ActionIDs.h"
 #include "input/actions/ActionTranslator.h"
 #include "input/keyboard/Key.h"
-#include "input/keyboard/KeyIDs.h"
 #include "input/keymaps/interfaces/IKeyMapper.h"
 #include "input/keymaps/joysticks/GamepadTranslator.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 #include "input/keymaps/keyboard/KeyboardTranslator.h"
 #include "input/keymaps/remote/CustomControllerTranslator.h"
 #include "input/keymaps/remote/IRTranslator.h"
