+++ xbmc/interfaces/swig/AddonModuleXbmcgui.i
@@ -21,7 +21,7 @@
 #include "interfaces/legacy/Dialog.h"
 #include "interfaces/legacy/WindowXML.h"
 #include "input/actions/ActionIDs.h"
-#include "input/keyboard/KeyIDs.h"
+#include "input/keymaps/keyboard/KeyIDs.h"
 
 using namespace XBMCAddon;
 using namespace xbmcgui;
@@ -108,4 +108,4 @@ using namespace xbmcgui;
 %include "interfaces/legacy/WindowXML.h"
 
 %include "input/actions/ActionIDs.h"
-%include "input/keyboard/KeyIDs.h"
+%include "input/keymaps/keyboard/KeyIDs.h"
