+++ xbmc/peripherals/devices/PeripheralCecAdapter.cpp
@@ -19,7 +19,7 @@
 #include "guilib/GUIWindowManager.h"
 #include "input/actions/Action.h"
 #include "input/actions/ActionIDs.h"
-#include "input/remote/IRRemote.h"
+#include "input/keymaps/remote/IRRemoteIDs.h"
 #include "interfaces/AnnouncementManager.h"
 #include "jobs/JobManager.h"
 #include "messaging/ApplicationMessenger.h"
